httpサーバ ポート番号取得処理を別ファイルに切り出した
